pandasai_mysql/requirements.txt
pandasai[connectors]
streamlit
langchain_anthropic
seaborn
python_dotenv

langchain_groq
